fix(swingset): change dispatch from object to function
Previously, the lowest level of a vat was defined by an object named
`dispatch`, which had three methods: `{ deliver, notify, dropExports }`. Now
we define `dispatch()` to be a *function*, which takes a
`VatDeliveryObject` (of which there are three flavors). This simplifies the
supervisor code and removes a lot of redundant boilerplate, and prepares for
later phase where `dispatch()` becomes `async dispatch()`.

refs #2671